requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
grpcio-reflection==1.71.0
protobuf==5.29.5
motor==3.6.0
pymongo==4.9.0  # Изменено с 4.13.2 на 4.9.0
pydantic==2.7.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
pyjwt==2.8.0
dishka==1.6.0
structlog==24.4.0
redis==5.0.8
aio-pika==9.4.0
pytest==8.3.2
pytest-asyncio==0.23.8
aiohttp==3.10.5
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-jose==3.3.0